chore: Update docblocks to satisfy phpstan 10, prevent edge case fatal errors on unsupported ID

+++ tests/Unit/Command/SyncTestTemplatesCommandTest.php
@@ -325,6 +325,10 @@ TWIG;
             \RecursiveIteratorIterator::CHILD_FIRST
         );
         foreach ($files as $file) {
+            if (!$file instanceof \SplFileInfo) {
+                continue;
+            }
+
             $file->isDir() ? rmdir($file->getPathname()) : unlink($file->getPathname());
         }
         rmdir($dir);
